Moved Qt translator implementation from tagger to i18n.qt
